ai_gym_backend/requirements.txt
# ── Web Framework ──────────────────────────────────────────────────────────────
fastapi==0.135.3
uvicorn==0.44.0
starlette==1.0.0
python-multipart==0.0.24

# ── Database ───────────────────────────────────────────────────────────────────
SQLAlchemy==2.0.49
greenlet==3.4.0

# ── AI / LLM ───────────────────────────────────────────────────────────────────
google-genai==1.72.0

# ── Vector Database (RAG) ──────────────────────────────────────────────────────
# NOTE: Install with --prefer-binary to avoid C++ compiler requirement on Windows
# pip install chromadb --prefer-binary
chromadb==1.5.7

# ── Machine Learning ───────────────────────────────────────────────────────────
scikit-learn==1.8.0
numpy==2.4.4
pandas==3.0.2
scipy==1.17.1
joblib==1.5.3
threadpoolctl==3.6.0

# ── Computer Vision ────────────────────────────────────────────────────────────
mediapipe==0.10.33
opencv-python==4.13.0.92
opencv-contrib-python==4.13.0.92
flatbuffers==25.12.19

# ── NLP / Sentiment Analysis ───────────────────────────────────────────────────
vaderSentiment==3.3.2
textblob==0.20.0
nltk==3.9.4
regex==2026.4.4

# ── Cloud Storage ──────────────────────────────────────────────────────────────
cloudinary==1.44.1

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv==1.2.2
Pillow==12.2.0
httpx==0.28.1
requests==2.33.1
pydantic==2.12.5
pydantic-settings==2.13.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
tqdm==4.67.3
rich==14.3.3
orjson==3.11.8
#IoT
paho-mqtt==2.1.0